# perf(frontend): 初期描画最適化（tw-animate-css削除＋LCP画像priority/sizes付与）