--- a/apps/frontend/app/globals.css
+++ b/apps/frontend/app/globals.css
@@ -1,5 +1,4 @@
 @import "tailwindcss";
-@import "tw-animate-css";
 
 @custom-variant dark (&:is(.dark *));