--- a/apps/frontend/package.json
+++ b/apps/frontend/package.json
@@ -50,7 +50,6 @@
     "tailwind-merge": "3.3.1",
     "tailwindcss": "4.1.13",
     "tiny-invariant": "1.3.3",
-    "tw-animate-css": "1.3.8",
     "zustand": "5.0.8"
   },
   "devDependencies": {