概要
- 開発ビルドでFCP悪化の一因だったグローバルCSSを整理し、本番初期描画の軽量化を維持しました。
- 本番ビルド（:33001）のSlow 4G + 4x 条件でも FCP/LCP ≈ 0.65s 台を確認。

変更点
- app/globals.css: `@import "tw-animate-css"` を削除（Tailwindコア＋`tailwindcss-animate`で代替）。
- components/restaurant-detail.tsx: メイン画像に `priority` と `sizes="(max-width: 768px) 100vw, 600px"` を付与。
- apps/frontend/package.json: 未使用依存 `tw-animate-css` を削除。

計測（:33001, 本番ビルド）
- /meshi/18: FCP ≈ 668 ms, LCP ≈ 666 ms（Slow 4G + 4x）, CLS 0.00
- /meshi/15: FCP ≈ 648 ms, LCP ≈ 645 ms（Slow 4G + 4x）, CLS 0.00

リスク/互換性
- `tw-animate-css` 固有ユーティリティの利用なしを全コード検索で確認。UI崩れの可能性は低い想定。

チェックリスト
- [ ] `pnpm format`
- [ ] `pnpm lint`
- [ ] 主要ページのE2Eスタイル崩れ確認